+++ integtest/spec/helper/dsl/convert_single.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require 'digest'
+
 module Dsl
   module ConvertSingle
     ##
@@ -24,12 +26,6 @@ module Dsl
     end
     shared_context 'convert single' do
       let(:out) { @asciidoctor_out }
-      let(:asciidoctor_files) do
-        files_in(dest_file('.')).reject { |f| f.start_with? 'asciidoc/' }
-      end
-      let(:asciidoc_files) do
-        files_in(dest_file('asciidoc'))
-      end
       it 'prints the path to the html index' do
         expect(out).to include(dest_file('index.html'))
       end
@@ -42,28 +38,79 @@ module Dsl
       it 'creates the js' do
         expect(dest_file('docs.js')).to file_exist
       end
-      it 'logs the same lines with asciidoc' do
-        # The only difference should be that the output path
-        # includes `asciidoc/`
-        expect(@asciidoc_out.gsub('asciidoc/', '')).to eq(@asciidoctor_out)
-      end
-      it 'makes the same files with asciidoc' do
-        expect(asciidoc_files).to eq(asciidoctor_files)
-      end
-      it 'makes exactly the same html files with asciidoc' do
-        # This does *all* the files in the same example which doesn't feel very
-        # rspec but it gets the job done and we don't know the file list at this
-        # point so there isn't much we can do about it.
-        asciidoctor_files.each do |file|
-          next unless File.extname(file) == '.html'
+      context 'when run with asciidoc' do
+        let(:asciidoctor_files) do
+          files_in(dest_file('.')).reject { |f| f.start_with? 'asciidoc/' }
+        end
+        let(:asciidoctor_non_snippet_files) do
+          asciidoctor_files.reject { |f| File.extname(f) == '.console' }
+        end
+        let(:asciidoctor_snippet_files) do
+          asciidoctor_files.select { |f| File.extname(f) == '.console' }
+        end
+        let(:asciidoc_files) do
+          files_in(dest_file('asciidoc'))
+        end
+        let(:asciidoc_non_snippet_files) do
+          asciidoc_files.reject { |f| File.extname(f) == '.json' }
+        end
+        let(:asciidoc_snippet_files) do
+          asciidoc_files.select { |f| File.extname(f) == '.json' }
+        end
+        it 'logs the same lines' do
+          # The only difference should be that the output path
+          # includes `asciidoc/`
+          expect(@asciidoc_out.gsub('asciidoc/', '')).to eq(@asciidoctor_out)
+        end
+        it 'makes the same non-snippet files' do
+          expect(asciidoc_non_snippet_files).to contain_exactly(
+            *asciidoctor_non_snippet_files
+          )
+        end
+        it 'makes *almost* the same html files' do
+          # This does *all* the files in the same example which doesn't feel
+          # very rspec but it gets the job done and we don't know the file list
+          # at this point so there isn't much we can do about it.
+          asciidoctor_files.each do |file|
+            next unless File.extname(file) == '.html'
 
-          # We shell out to the html_diff tool that we wrote for this when the
-          # integration tests were all defined in a Makefile. It isn't great to
-          # shell out here but we've already customized html_diff.
-          asciidoctor_file = dest_file(file)
-          asciidoc_file = dest_file("asciidoc/#{file}")
-          html_diff = File.expand_path '../../../html_diff', __dir__
-          sh "#{html_diff} #{asciidoctor_file} #{asciidoc_file}"
+            # We shell out to the html_diff tool that we wrote for this when
+            # the integration tests were all defined in a Makefile. It isn't
+            # great to shell out here but we've already customized html_diff.
+            asciidoctor_file = dest_file(file)
+            asciidoc_file = dest_file("asciidoc/#{file}")
+            html_diff = File.expand_path '../../../html_diff', __dir__
+            sh "#{html_diff} #{asciidoctor_file} #{asciidoc_file}"
+          end
+        end
+        it 'makes exactly the same non-snippet, non-html files' do
+          asciidoctor_non_snippet_files.each do |file|
+            next if File.extname(file) == '.html'
+
+            asciidoctor_digest = Digest::MD5.file dest_file(file)
+            asciidoc_digest = Digest::MD5.file dest_file("asciidoc/#{file}")
+            expect(asciidoc_digest).to eq(asciidoctor_digest)
+          end
+        end
+        it 'makes the same snippet files but with different names' do
+          asciidoc = {}
+          asciidoc_snippet_files.each do |file|
+            asciidoc_file = dest_file "asciidoc/#{file}"
+            snippet = File.open asciidoc_file, 'r:UTF-8', &:read
+            # Strip trailing spaces from the snippet because Asciidoctor
+            # does that and AsciiDoc does not.
+            snippet = snippet.lines.map(&:rstrip).join("\n")
+            # Add a trailing newline to the snippet because Asciidoctor
+            # does that and AsciiDoc does not.
+            snippet += "\n"
+            asciidoc[snippet] = file
+          end
+          asciidoctor = {}
+          asciidoctor_snippet_files.each do |file|
+            snippet = File.open(dest_file(file), 'r:UTF-8', &:read)
+            asciidoctor[snippet] = file
+          end
+          expect(asciidoc).to have_same_keys(asciidoctor)
         end
       end
     end
